synbiobot_DNA_size/requirements.txt
biopython==1.78
pyairtable==1.4.0
benchlingapi==2.1.12
validators==0.18.2

